Inflation is calculated from the Account class expected inflation

+++ FinancialWebApp/Models/DataPoints.cs
@@ -10,7 +10,7 @@ namespace FinancialWebApp.Models
     [DataContract]
     public class DataPoint
     {
-        public DataPoint(int x, int y)
+        public DataPoint(long x, long y)
         {
             this.X = x;
             this.Y = y;
@@ -18,10 +18,10 @@ namespace FinancialWebApp.Models
 
         //Explicitly setting the name to be used while serializing to JSON.
         [DataMember(Name = "x")]
-        public Nullable<int> X = null;
+        public Nullable<long> X = null;
 
         //Explicitly setting the name to be used while serializing to JSON.
         [DataMember(Name = "y")]
-        public Nullable<int> Y = null;
+        public Nullable<long> Y = null;
     }
 }
